EvSeg

Documentation for EvSeg.

EvSeg.min_dist

EvSeg.min_dist — Method

min_dist(V::Vector)::Real

Return the sum of all elements included in a vector .

Arguments

s1 – Vector of event boundaries (first segmentation of interest)

s2 – Vector of event boundaries (second segmentation of interest)

tstart – start of the video/segmented content

tend – end of the video/segmented content

Output

A real number representing the minimal distance computed between two segmentations of interest.

Notes

Make sure not to include NaNs or character vectors.

Examples

julia> rse_mean([1,5,6])
12

julia> rse_mean(1:36)
666

See also min_dist